Changed INT terminals from ecore::EInt to ecore::ELong
Should have no effect on generated platform code, as this is only used
internally in the ThingML compilers.

Fixes #132

+++ compilers/c/src/main/java/org/thingml/compilers/c/CCfgMainGenerator.java
@@ -1282,7 +1282,7 @@ public class CCfgMainGenerator extends CfgMainGenerator {
             builder.append("dispatch_" + m.getName() + "(");
             builder.append("(mbuf[0] << 8) + mbuf[1] /* instance port*/");
 
-            int idx = 2;
+            long idx = 2;
 
             for (Parameter pt : m.getParameters()) {
                 //builder.append(",\n" + ctx.deserializeFromByte(pt.getType(), "mbuf", idx, ctx) + " /* " + pt.getName() + " */ ");
