+++ compilers/c/src/main/java/org/thingml/compilers/c/posixmt/PosixMTThingImplCompiler.java
@@ -408,7 +408,7 @@ public class PosixMTThingImplCompiler extends CThingImplCompiler {
                         //builder.append("_instance");
                         builder.append("(_instance");
 
-                        int idx = 2;
+                        long idx = 2;
 
                         for (Parameter pt : m.getParameters()) {
                             //builder.append(",\n" + ctx.deserializeFromByte(pt.getType(), "mbuf", idx, ctx) + " /* " + pt.getName() + " */ ");
